# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  income: "income",
  expense: "expense",
} as const;

export interface Transaction {
  id: number;
  amount: number;
  type: TransactionType;
  description: string;
  date: string;
  categoryId?: number | null;
  categoryName?: string | null;
  categoryColor?: string | null;
  createdAt: string;
}

export type CreateTransactionInputType =
  (typeof CreateTransactionInputType)[keyof typeof CreateTransactionInputType];

export const CreateTransactionInputType = {
  income: "income",
  expense: "expense",
} as const;

export interface CreateTransactionInput {
  /** @minimum 0.01 */
  amount: number;
  type: CreateTransactionInputType;
  /** @minLength 1 */
  description: string;
  date: string;
  categoryId?: number | null;
}

export type CategoryType = (typeof CategoryType)[keyof typeof CategoryType];

export const CategoryType = {
  income: "income",
  expense: "expense",
  both: "both",
} as const;

export interface Category {
  id: number;
  name: string;
  color: string;
  icon?: string | null;
  type: CategoryType;
  createdAt: string;
}

export type CreateCategoryInputType =
  (typeof CreateCategoryInputType)[keyof typeof CreateCategoryInputType];

export const CreateCategoryInputType = {
  income: "income",
  expense: "expense",
  both: "both",
} as const;

export interface CreateCategoryInput {
  /** @minLength 1 */
  name: string;
  color: string;
  icon?: string | null;
  type: CreateCategoryInputType;
}

export interface Budget {
  id: number;
  categoryId: number;
  categoryName: string;
  categoryColor: string;
  /** YYYY-MM format */
  month: string;
  amount: number;
  spent: number;
  remaining: number;
  percentUsed: number;
}

export interface CreateBudgetInput {
  categoryId: number;
  /** YYYY-MM format */
  month: string;
  /** @minimum 0.01 */
  amount: number;
}

export interface MonthlySummary {
  month: string;
  totalIncome: number;
  totalExpenses: number;
  savings: number;
  savingsRate: number;
  transactionCount: number;
}

export interface CategorySummary {
  categoryId?: number | null;
  categoryName: string;
  categoryColor: string;
  total: number;
  count: number;
  percentage: number;
}

export interface MonthlyTrend {
  month: string;
  income: number;
  expenses: number;
  savings: number;
}

export type GetTransactionsParams = {
  /**
   * Filter by month in YYYY-MM format
   */
  month?: string;
  categoryId?: number;
  type?: GetTransactionsType;
};

export type GetTransactionsType =
  (typeof GetTransactionsType)[keyof typeof GetTransactionsType];

export const GetTransactionsType = {
  income: "income",
  expense: "expense",
} as const;

export type GetBudgetsParams = {
  /**
   * Filter by month in YYYY-MM format
   */
  month?: string;
};

export type GetMonthlySummaryParams = {
  /**
   * Month in YYYY-MM format
   */
  month?: string;
};

export type GetCategorySummaryParams = {
  /**
   * Month in YYYY-MM format
   */
  month?: string;
};