# lib/api-zod/src/generated/types/budget.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Budget {
  id: number;
  categoryId: number;
  categoryName: string;
  categoryColor: string;
  /** YYYY-MM format */
  month: string;
  amount: number;
  spent: number;
  remaining: number;
  percentUsed: number;
}