# lib/api-zod/src/generated/types/categorySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CategorySummary {
  categoryId?: number | null;
  categoryName: string;
  categoryColor: string;
  total: number;
  count: number;
  percentage: number;
}