# lib/api-zod/src/generated/types/createBudgetInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBudgetInput {
  categoryId: number;
  /** YYYY-MM format */
  month: string;
  /** @minimum 0.01 */
  amount: number;
}