# lib/api-zod/src/generated/types/createCategoryInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCategoryInputType } from "./createCategoryInputType";

export interface CreateCategoryInput {
  /** @minLength 1 */
  name: string;
  color: string;
  icon?: string | null;
  type: CreateCategoryInputType;
}