# lib/api-zod/src/generated/types/createTransactionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTransactionInputType } from "./createTransactionInputType";

export interface CreateTransactionInput {
  /** @minimum 0.01 */
  amount: number;
  type: CreateTransactionInputType;
  /** @minLength 1 */
  description: string;
  date: Date;
  categoryId?: number | null;
}